Refactor: diretory structure
