+++ src/App.js
@@ -1,14 +1,14 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import ScrollRestoration from "react-scroll-restoration";
-import Skeleton from "@components/Skeleton/Skeleton";
+import Skeleton from "components/Skeleton/Skeleton";
 
-import Login from "@components/Login/Login";
-import Search from "@components/Search/Search/Search";
-import AddRoom from "@components/Search/AddRoom/AddRoom";
-import Myroom from "@components/Room/Myroom/Myroom";
-import Mypage from "@components/Mypage/Mypage";
-import WrapChat from "@components/Chatting/WrapChat";
+import Login from "components/Login/Login";
+import Search from "components/Search/Search";
+import AddRoom from "components/AddRoom/AddRoom";
+import Myroom from "components/Myroom/Myroom";
+import Mypage from "components/Mypage/Mypage";
+import WrapChat from "components/Chatting/WrapChat";
 
 import "App.css";
 import "Font.css";
