+++ src/components/Chatting/MessagesBody/ChatSet.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import moment from "moment";
-import ProfileImg from "@components/Mypage/ProfileImg";
-import { getS3Url } from "@tools/trans";
+import ProfileImg from "components/Mypage/ProfileImg";
+import { getS3Url } from "tools/trans";
 import PropTypes from "prop-types";
 
 const ChatImageLoading = (props) => {
