+++ src/components/Mypage/PopupPolicy/PopupPolicy.jsx
@@ -3,7 +3,7 @@ import { animated, useSpring } from "react-spring";
 import { useHistory } from "react-router";
 import RLayout from "components/common/RLayout";
 import PropTypes from "prop-types";
-import axios from "@tools/axios";
+import axios from "tools/axios";
 
 import SparcsLogoBlack from "static/assets/SparcsLogoBlack.svg";
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
