+++ src/components/common/roomOptions/Place.jsx
